requirements.txt
imap-tools
google-generativeai
requests
python-dotenv
huggingface_hub
fpdf2
weasyprint
beautifulsoup4
